requirements_test.txt
django-user-accounts